breakdown asset modules
