+++ src/AvmAssets/index.ts
@@ -1,0 +1,4 @@
+export * from '../Erc20/Erc20';
+export * from './getAssetDescription';
+export * from '../Erc20/Erc20Token';
+export * from './types';
