+++ src/AvmAssets/types.ts
@@ -1,5 +1,4 @@
 import { Buffer } from 'avalanche';
-import Erc20Token from '@/Asset/Erc20Token';
 
 export interface iAssetCache {
     [assetId: string]: iAssetDescriptionClean;
@@ -17,16 +16,4 @@ export interface iAssetDescriptionClean {
     symbol: string;
     assetID: string;
     denomination: number;
-}
-
-export interface Erc20TokenData {
-    chainId: number;
-    address: string;
-    decimals: number;
-    name: string;
-    symbol: string;
-}
-
-export interface Erc20Store {
-    [address: string]: Erc20Token;
 }
