+++ src/Erc20/index.ts
@@ -1,4 +1,4 @@
 export * from './Erc20';
-export * from './Assets';
+export * from './erc20Cache';
 export * from './Erc20Token';
 export * from './types';
