+++ src/index.ts
@@ -1,12 +1,13 @@
 export * from '@/Keystore';
 export * from '@/Network';
-export * from '@/Asset';
+export * from '@/AvmAssets';
+export * from '@/Erc20';
 export * from '@/History';
 export * from '@/Wallet';
-
+export * from '@/Explorer';
+export * from '@/helpers';
+export * from '@/utils';
 export * from './types';
 
 export { BN } from 'avalanche';
 export { Big } from 'big.js';
-
-export * from '@/helpers';
